requirements.txt
Flask==3.1.0
Pillow==11.0.0
reportlab==4.2.5
gunicorn==23.0.0